+++ app/domain/User.js
@@ -4,7 +4,7 @@ import Profile from './Profile';
 
 export default class User {
 
-  @observable id: string;
+  id: string = '';
   @observable profile: Profile;
 
   constructor(id: string, profile: Profile) {