+++ app/domain/WalletAccount.js
@@ -1,17 +1,20 @@
 // @flow
 import { observable } from 'mobx';
 import BigNumber from 'bignumber.js';
+import WalletAddress from './WalletAddress';
 
-export default class WalletAddress {
+export default class WalletAccount {
 
   id: string = '';
+  @observable name: string = '';
   @observable amount: BigNumber;
-  @observable isUsed: boolean = false;
+  @observable addresses: Array<WalletAddress> = [];
 
   constructor(data: {
     id: string,
+    name: string,
     amount: BigNumber,
-    isUsed: boolean,
+    addresses: Array<WalletAddress>,
   }) {
     Object.assign(this, data);
   }