+++ app/stores/ada/AdaAddressesStore.js
@@ -1,6 +1,6 @@
 // @flow
 import { observable, computed, action, runInAction } from 'mobx';
-import _ from 'lodash';
+import { find } from 'lodash';
 import Store from '../lib/Store';
 import CachedRequest from '../lib/LocalizedCachedRequest';
 import Request from '../lib/LocalizedRequest';
@@ -8,7 +8,7 @@ import WalletAddress from '../../domain/WalletAddress';
 import LocalizableError from '../../i18n/LocalizableError';
 import type { GetAddressesResponse, CreateAddressResponse } from '../../api/ada/index';
 
-export default class AddressesStore extends Store {
+export default class AdaAddressesStore extends Store {
 
   @observable lastGeneratedAddress: ?WalletAddress = null;
   @observable addressesRequests: Array<{
@@ -97,7 +97,7 @@ export default class AddressesStore extends Store {
   };
 
   _getAddressesAllRequest = (walletId: string): CachedRequest<GetAddressesResponse> => {
-    const foundRequest = _.find(this.addressesRequests, { walletId });
+    const foundRequest = find(this.addressesRequests, { walletId });
     if (foundRequest && foundRequest.allRequest) return foundRequest.allRequest;
     return new CachedRequest(this.api.ada.getAddresses);
   };